moved lexer to its own module to clean up code and be more clear with where the lexer is.
